MINOR: Merge kafka.security classes into test-common
