+++ build.gradle
@@ -1113,9 +1113,6 @@ project(':core') {
     implementation.exclude module: 'jmxri'
     implementation.exclude module: 'jmxtools'
     implementation.exclude module: 'mail'
-    // To prevent a UniqueResourceException due the same resource existing in both
-    // org.apache.directory.api/api-all and org.apache.directory.api/api-ldap-schema-data
-    testImplementation.exclude module: 'api-ldap-schema-data'
     releaseOnly
   }
 
@@ -1169,22 +1166,7 @@ project(':core') {
     testImplementation project(':test-common:test-common-runtime')
     testImplementation project(':test-common:test-common-internal-api')
     testImplementation project(':test-common:test-common-util')
-    testImplementation libs.bcpkix
     testImplementation libs.mockitoCore
-    testImplementation(libs.apacheda) {
-      exclude group: 'xml-apis', module: 'xml-apis'
-      // `mina-core` is a transitive dependency for `apacheds` and `apacheda`.
-      // It is safer to use from `apacheds` since that is the implementation.
-      exclude module: 'mina-core'
-    }
-    testImplementation libs.apachedsCoreApi
-    testImplementation libs.apachedsInterceptorKerberos
-    testImplementation libs.apachedsProtocolShared
-    testImplementation libs.apachedsProtocolKerberos
-    testImplementation libs.apachedsProtocolLdap
-    testImplementation libs.apachedsLdifPartition
-    testImplementation libs.apachedsMavibotPartition
-    testImplementation libs.apachedsJdbmPartition
     testImplementation libs.junitJupiter
     testImplementation libs.caffeine
     testImplementation testLog4j2Libs
@@ -1684,6 +1666,16 @@ project(':test-common:test-common-runtime') {
     archivesName = "kafka-test-common-runtime"
   }
 
+  configurations {
+    // To prevent a UniqueResourceException due the same resource existing in both
+    // org.apache.directory.api/api-all and org.apache.directory.api/api-ldap-schema-data
+    implementation.exclude module: 'api-ldap-schema-data'
+    // apacheds transitively pulls in the old bcprov-jdk15on which clashes with the modern
+    // bcpkix-jdk18on (different artifact coordinates, so Gradle does not conflict-resolve
+    // them). Exclude the old one; bcprov-jdk18on from bcpkix is sufficient.
+    implementation.exclude group: 'org.bouncycastle', module: 'bcprov-jdk15on'
+  }
+
   dependencies {
     api project(':core')
     api project(':clients')
@@ -1700,6 +1692,21 @@ project(':test-common:test-common-runtime') {
     implementation libs.junitJupiter
     implementation libs.jacksonDataformatYaml
     implementation libs.slf4jApi
+    implementation libs.bcpkix
+    implementation(libs.apacheda) {
+      exclude group: 'xml-apis', module: 'xml-apis'
+      // `mina-core` is a transitive dependency for `apacheds` and `apacheda`.
+      // It is safer to use from `apacheds` since that is the implementation.
+      exclude module: 'mina-core'
+    }
+    implementation libs.apachedsCoreApi
+    implementation libs.apachedsInterceptorKerberos
+    implementation libs.apachedsProtocolShared
+    implementation libs.apachedsProtocolKerberos
+    implementation libs.apachedsProtocolLdap
+    implementation libs.apachedsLdifPartition
+    implementation libs.apachedsMavibotPartition
+    implementation libs.apachedsJdbmPartition
     implementation testFixtures(project(':clients'))
 
     testImplementation testFixtures(project(':clients'))
