+++ checkstyle/suppressions.xml
@@ -317,7 +317,4 @@
     <suppress checks="(ClassDataAbstractionCoupling|ClassFanOutComplexity)"
               files="(ReplicaFetcherThreadBenchmark).java"/>
 
-     <!-- Add the new suppression rule for JaasTestUtils.java -->
-    <suppress checks="ImportControl" files="(JaasTestUtils).java" />
-
 </suppressions>
