+++ core/src/test/scala/integration/kafka/api/GroupEndToEndAuthorizationTest.scala
@@ -17,10 +17,10 @@
 package kafka.api
 
 import kafka.api.GroupEndToEndAuthorizationTest._
-import kafka.security.JaasTestUtils
 import org.apache.kafka.common.config.internals.BrokerSecurityConfigs
 import org.apache.kafka.common.security.auth.{AuthenticationContext, KafkaPrincipal, SaslAuthenticationContext}
 import org.apache.kafka.common.security.authenticator.DefaultKafkaPrincipalBuilder
+import org.apache.kafka.common.test.JaasUtils
 
 object GroupEndToEndAuthorizationTest {
   val GroupPrincipalType = "Group"
@@ -29,7 +29,7 @@ object GroupEndToEndAuthorizationTest {
     override def build(context: AuthenticationContext): KafkaPrincipal = {
       context match {
         case ctx: SaslAuthenticationContext =>
-          if (ctx.server.getAuthorizationID == JaasTestUtils.KAFKA_SCRAM_USER)
+          if (ctx.server.getAuthorizationID == JaasUtils.KAFKA_SCRAM_USER)
             new KafkaPrincipal(GroupPrincipalType, ClientGroup)
           else
             new KafkaPrincipal(GroupPrincipalType, ctx.server.getAuthorizationID)
@@ -42,6 +42,6 @@ object GroupEndToEndAuthorizationTest {
 
 class GroupEndToEndAuthorizationTest extends SaslScramSslEndToEndAuthorizationTest {
   override val clientPrincipal = new KafkaPrincipal(GroupPrincipalType, ClientGroup)
-  override val kafkaPrincipal = new KafkaPrincipal(GroupPrincipalType, JaasTestUtils.KAFKA_SCRAM_ADMIN)
+  override val kafkaPrincipal = new KafkaPrincipal(GroupPrincipalType, JaasUtils.KAFKA_SCRAM_ADMIN)
   this.serverConfig.setProperty(BrokerSecurityConfigs.PRINCIPAL_BUILDER_CLASS_CONFIG, classOf[GroupPrincipalBuilder].getName)
 }
