+++ core/src/test/scala/integration/kafka/api/SaslMultiMechanismConsumerTest.scala
@@ -12,9 +12,9 @@
   */
 package kafka.api
 
-import kafka.security.JaasTestUtils
 import kafka.utils.{TestInfoUtils, TestUtils}
 import org.apache.kafka.common.security.auth.SecurityProtocol
+import org.apache.kafka.common.test.JaasUtils
 import org.junit.jupiter.api.{AfterEach, BeforeEach, TestInfo, Timeout}
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
@@ -31,7 +31,7 @@ class SaslMultiMechanismConsumerTest extends BaseConsumerTest with SaslSetup {
   @BeforeEach
   override def setUp(testInfo: TestInfo): Unit = {
     startSasl(jaasSections(kafkaServerSaslMechanisms, Some(kafkaClientSaslMechanism),
-      JaasTestUtils.KAFKA_SERVER_CONTEXT_NAME))
+      JaasUtils.KAFKA_SERVER_CONTEXT_NAME))
     super.setUp(testInfo)
   }
