+++ core/src/test/scala/integration/kafka/api/SslEndToEndAuthorizationTest.scala
@@ -17,14 +17,13 @@
 
 package kafka.api
 
-import kafka.security.JaasTestUtils
-
 import java.util.Properties
 import org.apache.kafka.common.config.SslConfigs
 import org.apache.kafka.common.config.internals.BrokerSecurityConfigs
 import org.apache.kafka.common.network.ConnectionMode
 import org.apache.kafka.common.security.auth._
 import org.apache.kafka.common.security.authenticator.DefaultKafkaPrincipalBuilder
+import org.apache.kafka.common.test.JaasUtils
 import org.junit.jupiter.api.{BeforeEach, TestInfo}
 
 import java.util.Optional
@@ -80,7 +79,7 @@ class SslEndToEndAuthorizationTest extends EndToEndAuthorizationTest {
   }
 
   override def clientSecurityProps(certAlias: String): Properties = {
-    val props = JaasTestUtils.securityConfigs(ConnectionMode.CLIENT, securityProtocol, OptionConverters.toJava(trustStoreFile),
+    val props = JaasUtils.securityConfigs(ConnectionMode.CLIENT, securityProtocol, OptionConverters.toJava(trustStoreFile),
       certAlias, clientCn, OptionConverters.toJava(clientSaslProperties), tlsProtocol, Optional.empty())
     props.remove(SslConfigs.SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG)
     props
@@ -88,7 +87,7 @@ class SslEndToEndAuthorizationTest extends EndToEndAuthorizationTest {
   // This test doesn't really care about matching the SSL certificate to a particular principal
   // We can override the CN and create a principal based on it or on the server SSL
   override def superuserSecurityProps(certAlias: String): Properties = {
-    val props = JaasTestUtils.securityConfigs(ConnectionMode.CLIENT, securityProtocol, OptionConverters.toJava(trustStoreFile),
+    val props = JaasUtils.securityConfigs(ConnectionMode.CLIENT, securityProtocol, OptionConverters.toJava(trustStoreFile),
       certAlias, superuserCn, OptionConverters.toJava(clientSaslProperties), tlsProtocol, Optional.empty())
     props.remove(SslConfigs.SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG)
     props
