+++ test-common/test-common-runtime/src/main/resources/minikdc.ldiff
